Working on modularizing worker
