+++ worker/pom.xml
@@ -24,12 +24,6 @@
 	  <artifactId>activemq-all</artifactId>
 	  <version>5.15.10</version>
 	</dependency>
-	<!-- https://mvnrepository.com/artifact/javax.jms/javax.jms-api -->
-	<dependency>
-	    <groupId>javax.jms</groupId>
-	    <artifactId>javax.jms-api</artifactId>
-	    <version>2.0.1</version>
-	</dependency>
 	<!-- DataBase Connector -->
 	<!-- https://mvnrepository.com/artifact/org.mariadb.jdbc/mariadb-java-client -->
 	<dependency>
@@ -149,6 +143,7 @@
 						    <element>PACKAGE</element>
 						    <excludes>
 						      <exclude>*worker</exclude>
+						      <exclude>*program*</exclude>
 						    </excludes>
 						    <limits>
 						      <limit>
@@ -162,6 +157,7 @@
 						    <element>CLASS</element>
 						    <excludes>
 						      <exclude>*Test</exclude>
+						      <exclude>*App</exclude>
 						      <exclude>*MariaDbWriter</exclude>
 						    </excludes>
 						    <limits>
@@ -252,7 +248,7 @@
 	                <!-- MainClass in mainfest make a executable jar -->
 					<archive>
 					  <manifest>
-						<mainClass>microtope.worker.App</mainClass>
+						<mainClass>microtope.program.App</mainClass>
 					  </manifest>
 					</archive>
 	            </configuration>
@@ -276,7 +272,7 @@
 			        <archive>
 			          <manifest>
 			            <addClasspath>true</addClasspath>
-			            <mainClass>microtope.worker.App</mainClass>
+			            <mainClass>microtope.program.App</mainClass>
 			          </manifest>
 			        </archive>
 			      </configuration>
