+++ worker/src/main/java/microtope/program/App.java
@@ -1,4 +1,4 @@
-package microtope.worker;
+package microtope.program;
 
 import java.io.IOException;
 import java.sql.SQLException;
@@ -8,6 +8,9 @@ import javax.jms.JMSException;
 
 import microtope.config.ActiveMqConfiguration;
 import microtope.config.SqlConfig;
+import microtope.worker.ActiveMqMessageReciever;
+import microtope.worker.DatabaseInsertListener;
+import microtope.worker.MariaDbWriter;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
